+++ src/event.rs
@@ -1,3 +1,0 @@
-enum AppEvent {
-  SshConnected(Result<SelectedHost>),
-}