+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use edit action instead of add, delete action
 - Move logic handle pressed key to `app.rs`
+- Upgrade UI with loading animation
 
 ### Fixed